@@ -1,10 +1,10 @@
-export interface CommonInputProps<T = string> {
+export type CommonInputProps<T = string> = {
   value: T;
   onChange: (newValue: T) => void;
   name?: string;
   placeholder?: string;
   disabled?: boolean;
-}
+};
 
 export enum IntegerVariant {
   UINT8 = "uint8",
