@@ -1,12 +1,11 @@
-import { useEffect } from "react";
 import { useEffectOnce, useLocalStorage, useReadLocalStorage } from "usehooks-ts";
+import { Chain, hardhat } from "viem/chains";
 import { Connector, useAccount, useConnect } from "wagmi";
-import { hardhat } from "wagmi/chains";
 import scaffoldConfig from "~~/scaffold.config";
-import { burnerWalletId, defaultBurnerChainId } from "~~/services/web3/wagmi-burner/BurnerConnector";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
+import { burnerWalletId } from "~~/services/web3/wagmi-burner/BurnerConnector";
+import { getTargetNetworks } from "~~/utils/scaffold-eth";
 
-const SCAFFOLD_WALLET_STROAGE_KEY = "scaffoldEth2.wallet";
+const SCAFFOLD_WALLET_STORAGE_KEY = "scaffoldEth2.wallet";
 const WAGMI_WALLET_STORAGE_KEY = "wagmi.wallet";
 
 // ID of the SAFE connector instance
@@ -14,11 +13,13 @@ const SAFE_ID = "safe";
 
 /**
  * This function will get the initial wallet connector (if any), the app will connect to
+ * @param initialNetwork
  * @param previousWalletId
  * @param connectors
  * @returns
  */
 const getInitialConnector = (
+  initialNetwork: Chain,
   previousWalletId: string,
   connectors: Connector[],
 ): { connector: Connector | undefined; chainId?: number } | undefined => {
@@ -29,14 +30,13 @@ const getInitialConnector = (
     return { connector: safeConnectorInstance };
   }
 
-  const targetNetwork = getTargetNetwork();
-  const allowBurner = scaffoldConfig.onlyLocalBurnerWallet ? targetNetwork.id === hardhat.id : true;
+  const allowBurner = scaffoldConfig.onlyLocalBurnerWallet ? initialNetwork.id === hardhat.id : true;
 
   if (!previousWalletId) {
     // The user was not connected to a wallet
     if (allowBurner && scaffoldConfig.walletAutoConnect) {
       const connector = connectors.find(f => f.id === burnerWalletId);
-      return { connector, chainId: defaultBurnerChainId };
+      return { connector, chainId: initialNetwork.id };
     }
   } else {
     // the user was connected to wallet
@@ -58,24 +58,22 @@ const getInitialConnector = (
  */
 export const useAutoConnect = (): void => {
   const wagmiWalletValue = useReadLocalStorage<string>(WAGMI_WALLET_STORAGE_KEY);
-  const [walletId, setWalletId] = useLocalStorage<string>(SCAFFOLD_WALLET_STROAGE_KEY, wagmiWalletValue ?? "");
+  const [walletId, setWalletId] = useLocalStorage<string>(SCAFFOLD_WALLET_STORAGE_KEY, wagmiWalletValue ?? "", {
+    initializeWithValue: false,
+  });
   const connectState = useConnect();
-  const accountState = useAccount();
-
-  useEffect(() => {
-    if (accountState.isConnected) {
-      // user is connected, set walletName
-      setWalletId(accountState.connector?.id ?? "");
-    } else {
-      // user has disconnected, reset walletName
+  useAccount({
+    onConnect({ connector }) {
+      setWalletId(connector?.id ?? "");
+    },
+    onDisconnect() {
       window.localStorage.setItem(WAGMI_WALLET_STORAGE_KEY, JSON.stringify(""));
       setWalletId("");
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [accountState.isConnected, accountState.connector?.name]);
+    },
+  });
 
   useEffectOnce(() => {
-    const initialConnector = getInitialConnector(walletId, connectState.connectors);
+    const initialConnector = getInitialConnector(getTargetNetworks()[0], walletId, connectState.connectors);
 
     if (initialConnector?.connector) {
       connectState.connect({ connector: initialConnector.connector, chainId: initialConnector.chainId });
