@@ -1,20 +1,23 @@
 import { useState } from "react";
+import { useTargetNetwork } from "./useTargetNetwork";
 import { Abi, ExtractAbiFunctionNames } from "abitype";
 import { useContractWrite, useNetwork } from "wagmi";
-import { getParsedError } from "~~/components/scaffold-eth";
 import { useDeployedContractInfo, useTransactor } from "~~/hooks/scaffold-eth";
-import { getTargetNetwork, notification } from "~~/utils/scaffold-eth";
+import { notification } from "~~/utils/scaffold-eth";
 import { ContractAbi, ContractName, UseScaffoldWriteConfig } from "~~/utils/scaffold-eth/contract";
 
 type UpdatedArgs = Parameters<ReturnType<typeof useContractWrite<Abi, string, undefined>>["writeAsync"]>[0];
 
 /**
- * @dev wrapper for wagmi's useContractWrite hook(with config prepared by usePrepareContractWrite hook) which loads in deployed contract abi and address automatically
+ * Wrapper around wagmi's useContractWrite hook which automatically loads (by name) the contract ABI and address from
+ * the contracts present in deployedContracts.ts & externalContracts.ts corresponding to targetNetworks configured in scaffold.config.ts
  * @param config - The config settings, including extra wagmi configuration
- * @param config.contractName - deployed contract name
+ * @param config.contractName - contract name
  * @param config.functionName - name of the function to be called
  * @param config.args - arguments for the function
  * @param config.value - value in ETH that will be sent with transaction
+ * @param config.blockConfirmations - number of block confirmations to wait for (default: 1)
+ * @param config.onBlockConfirmation - callback that will be called after blockConfirmations.
  */
 export const useScaffoldContractWrite = <
   TContractName extends ContractName,
@@ -32,10 +35,10 @@ export const useScaffoldContractWrite = <
   const { chain } = useNetwork();
   const writeTx = useTransactor();
   const [isMining, setIsMining] = useState(false);
-  const configuredNetwork = getTargetNetwork();
+  const { targetNetwork } = useTargetNetwork();
 
   const wagmiContractWrite = useContractWrite({
-    chainId: configuredNetwork.id,
+    chainId: targetNetwork.id,
     address: deployedContractData?.address,
     abi: deployedContractData?.abi as Abi,
     functionName: functionName as any,
@@ -60,7 +63,7 @@ export const useScaffoldContractWrite = <
       notification.error("Please connect your wallet");
       return;
     }
-    if (chain?.id !== configuredNetwork.id) {
+    if (chain?.id !== targetNetwork.id) {
       notification.error("You are on the wrong network");
       return;
     }
@@ -68,7 +71,7 @@ export const useScaffoldContractWrite = <
     if (wagmiContractWrite.writeAsync) {
       try {
         setIsMining(true);
-        await writeTx(
+        const writeTxResult = await writeTx(
           () =>
             wagmiContractWrite.writeAsync({
               args: newArgs ?? args,
@@ -77,9 +80,10 @@ export const useScaffoldContractWrite = <
             }),
           { onBlockConfirmation, blockConfirmations },
         );
+
+        return writeTxResult;
       } catch (e: any) {
-        const message = getParsedError(e);
-        notification.error(message);
+        throw e;
       } finally {
         setIsMining(false);
       }
