@@ -1,8 +1,12 @@
 import { useEffect, useMemo, useState } from "react";
+import { useTargetNetwork } from "./useTargetNetwork";
 import { Abi, AbiEvent, ExtractAbiEventNames } from "abitype";
+import { useInterval } from "usehooks-ts";
 import { Hash } from "viem";
+import * as chains from "viem/chains";
 import { usePublicClient } from "wagmi";
 import { useDeployedContractInfo } from "~~/hooks/scaffold-eth";
+import scaffoldConfig from "~~/scaffold.config";
 import { replacer } from "~~/utils/scaffold-eth/common";
 import {
   ContractAbi,
@@ -12,7 +16,7 @@ import {
 } from "~~/utils/scaffold-eth/contract";
 
 /**
- * @dev reads events from a deployed contract
+ * Reads events from a deployed contract
  * @param config - The config settings
  * @param config.contractName - deployed contract name
  * @param config.eventName - name of the event to listen for
@@ -21,6 +25,8 @@ import {
  * @param config.blockData - if set to true it will return the block data for each event (default: false)
  * @param config.transactionData - if set to true it will return the transaction data for each event (default: false)
  * @param config.receiptData - if set to true it will return the receipt data for each event (default: false)
+ * @param config.watch - if set to true, the events will be updated every pollingInterval milliseconds set at scaffoldConfig (default: false)
+ * @param config.enabled - set this to false to disable the hook from running (default: true)
  */
 export const useScaffoldEventHistory = <
   TContractName extends ContractName,
@@ -36,30 +42,45 @@ export const useScaffoldEventHistory = <
   blockData,
   transactionData,
   receiptData,
+  watch,
+  enabled = true,
 }: UseScaffoldEventHistoryConfig<TContractName, TEventName, TBlockData, TTransactionData, TReceiptData>) => {
   const [events, setEvents] = useState<any[]>();
-  const [isLoading, setIsLoading] = useState(true);
+  const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string>();
+  const [fromBlockUpdated, setFromBlockUpdated] = useState<bigint>(fromBlock);
+
   const { data: deployedContractData, isLoading: deployedContractLoading } = useDeployedContractInfo(contractName);
   const publicClient = usePublicClient();
+  const { targetNetwork } = useTargetNetwork();
 
-  useEffect(() => {
-    async function readEvents() {
-      try {
-        if (!deployedContractData) {
-          throw new Error("Contract not found");
-        }
+  const readEvents = async (fromBlock?: bigint) => {
+    setIsLoading(true);
+    try {
+      if (!deployedContractData) {
+        throw new Error("Contract not found");
+      }
 
-        const event = (deployedContractData.abi as Abi).find(
-          part => part.type === "event" && part.name === eventName,
-        ) as AbiEvent;
+      if (!enabled) {
+        throw new Error("Hook disabled");
+      }
 
+      const event = (deployedContractData.abi as Abi).find(
+        part => part.type === "event" && part.name === eventName,
+      ) as AbiEvent;
+
+      const blockNumber = await publicClient.getBlockNumber({ cacheTime: 0 });
+
+      if ((fromBlock && blockNumber >= fromBlock) || blockNumber >= fromBlockUpdated) {
         const logs = await publicClient.getLogs({
           address: deployedContractData?.address,
           event,
           args: filters as any, // TODO: check if it works and fix type
-          fromBlock,
+          fromBlock: fromBlock || fromBlockUpdated,
+          toBlock: blockNumber,
         });
+        setFromBlockUpdated(blockNumber + 1n);
+
         const newEvents = [];
         for (let i = logs.length - 1; i >= 0; i--) {
           newEvents.push({
@@ -79,23 +100,34 @@ export const useScaffoldEventHistory = <
                 : null,
           });
         }
-        setEvents(newEvents);
+        if (events && typeof fromBlock === "undefined") {
+          setEvents([...newEvents, ...events]);
+        } else {
+          setEvents(newEvents);
+        }
         setError(undefined);
-      } catch (e: any) {
-        console.error(e);
-        setEvents(undefined);
-        setError(e);
-      } finally {
-        setIsLoading(false);
       }
+    } catch (e: any) {
+      console.error(e);
+      setEvents(undefined);
+      setError(e);
+    } finally {
+      setIsLoading(false);
     }
+  };
+
+  useEffect(() => {
+    readEvents(fromBlock);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [fromBlock, enabled]);
+
+  useEffect(() => {
     if (!deployedContractLoading) {
       readEvents();
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     publicClient,
-    fromBlock,
     contractName,
     eventName,
     deployedContractLoading,
@@ -107,6 +139,22 @@ export const useScaffoldEventHistory = <
     transactionData,
     receiptData,
   ]);
+
+  useEffect(() => {
+    // Reset the internal state when target network or fromBlock changed
+    setEvents([]);
+    setFromBlockUpdated(fromBlock);
+    setError(undefined);
+  }, [fromBlock, targetNetwork.id]);
+
+  useInterval(
+    async () => {
+      if (!deployedContractLoading) {
+        readEvents();
+      }
+    },
+    watch ? (targetNetwork.id !== chains.hardhat.id ? scaffoldConfig.pollingInterval : 4_000) : null,
+  );
 
   const eventHistoryData = useMemo(
     () =>
