@@ -1,13 +1,15 @@
 import "abitype";
 
+type AddressType = string;
+
 declare module "viem/node_modules/abitype" {
   export interface Config {
-    AddressType: string;
+    AddressType: AddressType;
   }
 }
 
 declare module "abitype" {
   export interface Config {
-    AddressType: string;
+    AddressType: AddressType;
   }
 }
