@@ -1,7 +1,7 @@
-import * as chains from "wagmi/chains";
+import * as chains from "viem/chains";
 import scaffoldConfig from "~~/scaffold.config";
 
-export type TChainAttributes = {
+type ChainAttributes = {
   // color | [lightThemeColor, darkThemeColor]
   color: string | [string, string];
   // Used to fetch price by providing mainnet token address
@@ -9,7 +9,9 @@ export type TChainAttributes = {
   nativeCurrencyTokenAddress?: string;
 };
 
-export const NETWORKS_EXTRA_DATA: Record<string, TChainAttributes> = {
+export type ChainWithAttributes = chains.Chain & Partial<ChainAttributes>;
+
+export const NETWORKS_EXTRA_DATA: Record<string, ChainAttributes> = {
   [chains.hardhat.id]: {
     color: "#b8af0c",
   },
@@ -18,9 +20,6 @@ export const NETWORKS_EXTRA_DATA: Record<string, TChainAttributes> = {
   },
   [chains.sepolia.id]: {
     color: ["#5f4bb6", "#87ff65"],
-  },
-  [chains.goerli.id]: {
-    color: "#0975F6",
   },
   [chains.gnosis.id]: {
     color: "#48a9a6",
@@ -33,13 +32,13 @@ export const NETWORKS_EXTRA_DATA: Record<string, TChainAttributes> = {
     color: "#92D9FA",
     nativeCurrencyTokenAddress: "0x7D1AfA7B718fb893dB30A3aBc0Cfc608AaCfeBB0",
   },
-  [chains.optimismGoerli.id]: {
+  [chains.optimismSepolia.id]: {
     color: "#f01a37",
   },
   [chains.optimism.id]: {
     color: "#f01a37",
   },
-  [chains.arbitrumGoerli.id]: {
+  [chains.arbitrumSepolia.id]: {
     color: "#28a0f0",
   },
   [chains.arbitrum.id]: {
@@ -52,15 +51,12 @@ export const NETWORKS_EXTRA_DATA: Record<string, TChainAttributes> = {
     color: "#1969ff",
   },
   [chains.scrollSepolia.id]: {
-    color: '#fbebd4'
-  }
+    color: "#fbebd4",
+  },
 };
 
 /**
- * Gives the block explorer transaction URL.
- * @param network
- * @param txnHash
- * @dev returns empty string if the network is localChain
+ * Gives the block explorer transaction URL, returns empty string if the network is a local chain
  */
 export function getBlockExplorerTxLink(chainId: number, txnHash: string) {
   const chainNames = Object.keys(chains);
@@ -86,10 +82,8 @@ export function getBlockExplorerTxLink(chainId: number, txnHash: string) {
 }
 
 /**
- * Gives the block explorer Address URL.
- * @param network - wagmi chain object
- * @param address
- * @returns block explorer address URL and etherscan URL if block explorer URL is not present for wagmi network
+ * Gives the block explorer URL for a given address.
+ * Defaults to Etherscan if no (wagmi) block explorer is configured for the network.
  */
 export function getBlockExplorerAddressLink(network: chains.Chain, address: string) {
   const blockExplorerBaseURL = network.blockExplorers?.default?.url;
@@ -105,14 +99,11 @@ export function getBlockExplorerAddressLink(network: chains.Chain, address: stri
 }
 
 /**
- * @returns targetNetwork object consisting targetNetwork from scaffold.config and extra network metadata
+ * @returns targetNetworks array containing networks configured in scaffold.config including extra network metadata
  */
-
-export function getTargetNetwork(): chains.Chain & Partial<TChainAttributes> {
-  const configuredNetwork = scaffoldConfig.targetNetwork;
-
-  return {
-    ...configuredNetwork,
-    ...NETWORKS_EXTRA_DATA[configuredNetwork.id],
-  };
+export function getTargetNetworks(): ChainWithAttributes[] {
+  return scaffoldConfig.targetNetworks.map(targetNetwork => ({
+    ...targetNetwork,
+    ...NETWORKS_EXTRA_DATA[targetNetwork.id],
+  }));
 }
